prepare webapp
