+++ webapp/templates/index.html
@@ -8,7 +8,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
 
   <!-- Site Properties -->
-  <title>Fixed Menu Example - Semantic</title>
+  <title>freeSessionNet</title>
 
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/fomantic-ui@2.9.2/dist/semantic.min.css">
   <script src="https://cdn.jsdelivr.net/npm/fomantic-ui@2.9.2/dist/semantic.min.js"></script>
@@ -20,6 +20,11 @@
   .ui.menu .item img.logo {
     margin-right: 1.5em;
   }
+  img.logo{
+    height: 25em;
+    width: auto;
+  }
+
   .main.container {
     margin-top: 8em;
   }
@@ -42,8 +47,7 @@
   <div class="ui inverted menu">
     <div class="ui container">
       <a href="#" class="header item">
-        <img class="logo" src="./images/fsn.jpeg">
-        freeSessionNet
+        <h1>freeSessionNet</h1>
       </a>
     </div>
   </div>
@@ -51,6 +55,13 @@
   <div class="ui main text container">
     <h1 class="ui header">Durchsuche die Daten deiner Stadt</h1>
     <form class="ui action left icon input" method="POST">
+      <!--
+      <select class="ui search dropdown">
+        <option value="Dessau">Dessau</option>
+        <option value="Magdeburg">Magdeburg</option>
+        <option value="Magdeburg">Halle</option>
+      </select>
+      -->
       <i class="search icon"></i>
       <input name="in_search" type="text" placeholder="Suchen...">
       <button class="ui teal button" type="submit">Suchen</button>
@@ -60,19 +71,20 @@
 
   <div class="ui main text container">
     <h2 class="ui header">Suchergebnisse</h2>
+    <div class="description">{{ count_found }}</div>
     <div class="ui relaxed divided list">
       {% for item in items %}
         <div class="item">
           <i class="large angle right middle aligned icon"></i>
           <div class="content">
             <a href = {{ item['link'] }} target="_blank" >{{ item['link'] }}</a>
-            <div class="description">{{ item['resp'] }}</div>
+            <div class="description"><h3>{{ item['resp'] }}</h3></div>
             <div class="description">{{ item['desc'] }}</div>
-            <div class="description">{{ item['date'] }}</div>
+            <div class="description"><h4>{{ item['date'] }}</h4></div>
             <div class="description">{{ item['accuracy'] }}</div>
           </div>
         </div>
-      {% endfor %}  
+      {% endfor %}
     </div>
   </div>
 
@@ -80,7 +92,7 @@
 
   <div class="ui inverted vertical footer segment">
     <div class="ui center aligned container">
-      <img src="assets/images/logo.png" class="ui centered mini image">
+      <img class="ui centered image logo" src="{{url_for('static', filename='fsn.jpeg')}}">
       <div class="ui horizontal inverted small divided link list">
         <a class="item" href="#">Site Map</a>
         <a class="item" href="#">Contact Us</a>
